Recipes and Songs DB rework

+++ src/main/java/cz/gattserver/grass/recipes/model/RecipeRepositoryCustom.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface RecipeRepositoryCustom {
 
-	int count(String filter);
+	long count(String filter);
 
 	List<Recipe> fetch(String filter, int offset, int limit);
 }
